changed M_PI definition to ward off long double problems

+++ src/common/standard.h
@@ -28,7 +28,7 @@ typedef struct {
 
 #ifndef  PI
 #ifdef	M_PI
-#define	 PI		M_PI
+#define	 PI		((double)M_PI)
 #else
 #define	 PI		3.14159265358979323846
 #endif
